+++ build.gradle
@@ -32,7 +32,7 @@ javafx {
 
 application {
     // Change this to your main class.
-    mainClassName = "Launcher"
+    mainClassName = "duke.Launcher"
 }
 
 run {
