+++ data/duke.txt
@@ -1,6 +1,2 @@
+[E][✘] aba (at: 10/08/1998)
 [T][✘] this
-[T][✘] asjhaabkjasbf
-[T][✘] afknmaf
-[T][✘] asjrfhjafjew
-[T][✘] jabfkqwenkn qan
-[T][✘] jkasnfkjans
