+++ src/main/java/Launcher.java
@@ -1,3 +1,4 @@
+import duke.main.Main;
 import javafx.application.Application;
 
 public class Launcher {
