+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,9 @@
+package duke.command;
+
+public interface Command {
+
+    public static String run() {
+        return "";
+    }
+
+}
