+++ src/main/java/duke/command/ListCommand.java
@@ -1,22 +1,13 @@
-import java.util.Scanner;
+package duke.command;
 
-public class Ui {
-
-    private TaskList tasks;
-
-    /**
-     * Constructor for Ui class to run UI events
-      * @param tasks is the tasks that this current instance of Duke has
-     */
-    public Ui(TaskList tasks) {
-        this.tasks = tasks;
-    }
+import duke.main.TaskList;
 
 
+public class ListCommand implements Command {
     /**
      * Prints all the tasks in the current list
      */
-    public String printList() {
+    public static String run(TaskList tasks) {
         StringBuilder output = new StringBuilder();
         if (!tasks.isEmpty()) {
             output.append("Here are the tasks in your list" + "\n");
@@ -31,15 +22,7 @@ public class Ui {
         return output.toString();
     }
 
-
-    /**
-     * Prints the basic goodbye message when user has no more input
-     */
-    public void printGoodbye() {
-        System.out.println("GOODBYE!! MUAHAHHAHAHAHHAAHAHHAHAHA");
-    }
-
-
-
-
 }
+
+
+
