+++ src/main/java/duke/main/DialogBox.java
@@ -1,3 +1,5 @@
+package duke.main;
+
 import java.io.IOException;
 import java.util.Collections;
