+++ src/main/java/duke/main/Main.java
@@ -1,3 +1,4 @@
+package duke.main;
 
 import java.io.IOException;
 
@@ -8,7 +9,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
 /**
- * A GUI for Duke using FXML.
+ * A GUI for duke.main.Duke using FXML.
  */
 public class Main extends Application {
