+++ src/main/java/duke/main/MainWindow.java
@@ -1,3 +1,5 @@
+package duke.main;
+
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
@@ -9,7 +11,7 @@ import javafx.scene.layout.VBox;
 import java.io.IOException;
 
 /**
- * Controller for MainWindow. Provides the layout for the other controls.
+ * Controller for duke.main.MainWindow. Provides the layout for the other controls.
  */
 public class MainWindow extends AnchorPane {
     @FXML
@@ -30,7 +32,7 @@ public class MainWindow extends AnchorPane {
     public void initialize() {
         scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
         dialogContainer.getChildren().add(
-                DialogBox.getDukeDialog("Hello and welcome to Duke, how can I help?", dukeImage)
+                DialogBox.getDukeDialog("Hello and welcome to duke.main.Duke, how can I help?", dukeImage)
         );
     }
 
@@ -39,7 +41,7 @@ public class MainWindow extends AnchorPane {
     }
 
     /**
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * Creates two dialog boxes, one echoing user input and the other containing duke.main.Duke's reply and then appends them to
      * the dialog container. Clears the user input after processing.
      */
     @FXML
