+++ src/main/java/duke/main/Storage.java
@@ -1,3 +1,10 @@
+package duke.main;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -28,22 +35,27 @@ public class Storage {
             String desc = "";
             String at = "";
             String by = "";
-            if (type.equals("[T]")) {
-                desc = taskString.substring(taskString.indexOf(" ") + 1);
-                Todo todo = new Todo(desc);
-                tasks.add(todo);
-            } else if (type.equals("[D]")) {
-                desc = taskString.substring(taskString.indexOf(" ") + 1, taskString.indexOf(" (by"));
-                by = taskString.substring(taskString.indexOf("(by:") + 5, taskString.length() - 1);
-                LocalDate d = LocalDate.parse(by);
-                d.format(DateTimeFormatter.ofPattern("yyyy-mm-dd"));
-                Deadline deadline = new Deadline(desc, d.toString());
-                tasks.add(deadline);
-            } else if (type.equals("[E]")) {
-                desc = taskString.substring(taskString.indexOf(" ") + 1, taskString.indexOf(" (at"));
-                at = taskString.substring(taskString.indexOf("(at:") + 5, taskString.length() - 1);
-                Event event = new Event(desc, at);
-                tasks.add(event);
+            switch (type) {
+                case "[T]":
+                    desc = taskString.substring(taskString.indexOf(" ") + 1);
+                    Todo todo = new Todo(desc);
+                    tasks.add(todo);
+                    break;
+                case "[D]":
+                    desc = taskString.substring(taskString.indexOf(" ") + 1, taskString.indexOf(" (by"));
+                    by = taskString.substring(taskString.indexOf("(by:") + 5, taskString.length() - 1);
+                    LocalDate d = LocalDate.parse(by);
+                    d.format(DateTimeFormatter.ofPattern("yyyy-mm-dd"));
+                    Deadline deadline = new Deadline(desc, d.toString());
+                    tasks.add(deadline);
+                    break;
+                case "[E]":
+                    desc = taskString.substring(taskString.indexOf(" ") + 1, taskString.indexOf(" (at"));
+                    at = taskString.substring(taskString.indexOf("(at:") + 5, taskString.length() - 1);
+                    Event event = new Event(desc, at);
+                    tasks.add(event);
+                    break;
+
             }
 
         }
