+++ src/main/java/duke/main/Ui.java
@@ -1,0 +1,29 @@
+package duke.main;
+
+public class Ui {
+
+    private TaskList tasks;
+
+    /**
+     * Constructor for duke.main.Ui class to run UI events
+      * @param tasks is the tasks that this current instance of duke.main.Duke has
+     */
+    public Ui(TaskList tasks) {
+        this.tasks = tasks;
+    }
+
+
+
+
+
+    /**
+     * Prints the basic goodbye message when user has no more input
+     */
+    public void printGoodbye() {
+        System.out.println("GOODBYE!! MUAHAHHAHAHAHHAAHAHHAHAHA");
+    }
+
+
+
+
+}
