+++ src/test/java/EventTest.java
@@ -1,6 +1,5 @@
+import duke.task.Event;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
