+++ src/test/java/TodoTest.java
@@ -1,6 +1,5 @@
+import duke.task.Todo;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
